requirements.txt
streamlit
pandas
requests
tqdm
pymupdf
python-docx
groq
beautifulsoup4
lxml
xlsxwriter
openpyxl
tenacity
